Fixed warnings under WindRiver and updated WindRiver build targets ; replaced include in graphhost.h of stdint.h with types/vxTypes.h ; removed atomics from PIDController class

+++ KalmanGyro/KOP_ADXL345.cpp
@@ -41,7 +41,7 @@ KOP_ADXL345::~KOP_ADXL345() {
 // Actually measures y-axis of gyro for consistency with accelerometer
 int KOP_ADXL345::readGyroX() {
     if ( m_gyro != NULL ) {
-        return (m_gyro->GetAverageVoltage( m_gyroChannel ) - getGyroXzero()) / getGyroLSBsPerUnit();
+        return static_cast<int>((m_gyro->GetAverageVoltage( m_gyroChannel ) - getGyroXzero()) / getGyroLSBsPerUnit());
     }
     else {
         return 0;
